+++ CHANGES.rst
@@ -21,6 +21,10 @@ Improvements:
  * Privacy: Allow email registration when no IS (#2657).
  * Privacy: Settings: Add a Discovery section (#2606).
  * Privacy: Make NSContactsUsageDescription more generic and mention that 3pids are now uploaded hashed (#2521).
+ * Privacy: Settings: Add IDENTITY SERVER section (#2604).
+
+Bug fix:
+ * Theme: Make button theming work (#2734).
 
 Changes in 0.9.2 (2019-08-08)
 ===============================================